+++ agentpy/lists.py
@@ -113,16 +113,25 @@ class ObjList(list):
     def select(self, selection):
         """ Returns a new :class:`AgentList` based on `selection`.
 
-        Attributes:
+        Arguments:
             selection (list of bool): List with same length as the agent list.
                 Positions that return True will be selected.
         """
         return AgentList([a for a, s in zip(self, selection) if s])
 
-    def random(self, n=1):
-        """ Returns a new :class:`AgentList`
-        with ``n`` random agents (default 1)."""
-        return AgentList(rd.sample(self, n))
+    def random(self, n=1, generator=None):
+        """ Returns a new :class:`AgentList` with a random subset of agents.
+
+        Arguments:
+            n (int, optional): Number of agents (default 1).
+            generator (random.Random, optional): Random number generator.
+                 If none is passed, the hidden instance of :module:`random`
+                is used.
+        """
+        if generator:
+            return AgentList(generator.sample(self, n))
+        else:
+            return AgentList(rd.sample(self, n))
 
     def sort(self, var_key, reverse=False):
         """ Sorts the list based on var_key and returns self """